+++ src/typedefs.ts
@@ -1,4 +1,6 @@
 // https://www.typescriptlang.org/docs/handbook/utility-types.html
+import type { Gradient } from './gradient/gradient.class';
+import type { Pattern } from './pattern.class';
 
 interface NominalTag<T> {
   nominalTag?: T;
@@ -6,11 +8,19 @@ interface NominalTag<T> {
 
 type Nominal<Type, Tag> = NominalTag<Tag> & Type;
 
+// eslint-disable-next-line @typescript-eslint/ban-types
+type TNonFunctionPropertyNames<T> = {
+  [K in keyof T]: T[K] extends Function ? never : K;
+}[keyof T];
+export type TClassProperties<T> = Pick<T, TNonFunctionPropertyNames<T>>;
+
 const enum Degree {}
 const enum Radian {}
 
 export type TDegree = Nominal<number, Degree>;
 export type TRadian = Nominal<number, Radian>;
+
+export type TFiller = Gradient<'linear'> | Gradient<'radial'> | Pattern;
 
 export type TSize = {
   width: number;